# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyClinic API
  version: 1.0.0
  description: |2

    # INTRODUCTION

    ## Overview

    This app handles patient appointments.

    - For authentication see the [ACL AND USER MANAGEMENT](#tag/user-management) tag
    - To manage doctors availability see [Resource Management](#tag/resource-management)
    - Full booking flow is documented under [Appointments](#tag/appointments)

    ## System Requirements

    - Small clinic with 5 doctors
    - User to book a 30min appointment slot
    - Patient be able to cancel an appointment
    - Unbooked/cancelled slots to be available for any other booking

    ## Design Specifications

    These are the required specific functionalities derived from the basic requirements

    ### Public Side system

    - This is the client/patient side accessed for booking of appointments
    - Have a Registration/Login/Authentication system, for the user of type 'public' -- role based authentication
    - Be able to view all upcoming (self) appointments or bookings -- based on the logged in user
    - Can view all available time slots for a specific doctor in a given date -- so all slots sohuld be by date, and by doctor; or by doctor then by date
    - Can view and search for a doctor, then see their available slots for booking
    - Be able to book appointments (at least 30 mins after now)
    - Be able to cancel an appointment that has not yet arrived -- cancelling resets the slot to available
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...


...

...
paths:
  /api/acl/users:
    get:
      operationId: acl_users_list
      description: Base class view for user managemet
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - acl
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserListList'
          description: ''
    post:
      operationId: acl_users_create
      description: Create a user
      tags:
      - acl
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUp'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SignUp'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SignUp'
        required: true
      security:
      - BearerAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
          description: ''
...
components:
  schemas:
...
    BlankEnum:
      enum:
      - ''
    GenderEnum:
      enum:
      - male
      - female
      - other
      type: string
      description: |-
        * `male` - Male
        * `female` - Female
        * `other` - Rather Not say
...
    NullEnum:
      enum:
      - null
...
    PaginatedUserListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/UserList'
...
    SignUp:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        password:
          type: string
        username:
          type: string
          format: email
        other_names:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        residence:
          type: string
          nullable: true
        user_type:
          type: string
        date_of_birth:
          type: string
          format: date
          nullable: true
      required:
      - first_name
      - last_name
      - password
      - user_type
      - username
...
    UserList:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        other_names:
          type: string
          nullable: true
          maxLength: 255
        user_type:
          $ref: '#/components/schemas/UserTypeEnum'
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
      required:
      - first_name
      - last_name
      - username
    UserTypeEnum:
      enum:
      - admin
      - doctor
      - public
      type: string
      description: |-
        * `admin` - Admin
        * `doctor` - Doctor
        * `public` - public
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter JWT token as: Bearer <token>'
...